+++ src/utils/key-codes.js
@@ -1,0 +1,14 @@
+/*
+ * Key Codes (events)
+ */
+
+export default {
+  ENTER: 13,
+  LEFT: 37,
+  UP: 38,
+  RIGHT: 39,
+  DOWN: 40,
+  HOME: 36,
+  END: 35,
+  TAB: 9
+}